requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
pymongo==4.6.0
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
# passlib==1.7.4